Added seed-tests folder tests and test utils file.
